Rename float helpers
Change from `pullStart` to `pullLeft`, etc., because the 'start'/'end'
semantics only make sense if you're going to be hand-rolling an RTL
stylesheet. When relying on tools to automate the RTL flipping – like
it's clearer to keep the 'left' and 'right' words in the classname.

If you're authoring the primary stylesheet for RTL, then this change is
also an improvement. The `pullStart` class will always make an element
float left, which isn't the start of a RTL reading context. Instead,
`pullRight` would at least make sense for either RTL or LTR authoring.

Fix gh-1